changed obs space to float16

+++ stock_env.py
@@ -54,7 +54,7 @@ class StockEnv(Env):
         self.window_days = 5
         # Number of candles by number of featuers
         # TODO add previous action and holding time to state
-        self.observation_space = Box(low=0, high=np.inf, shape=(self.window_days*390,25), dtype=np.float64)
+        self.observation_space = Box(low=0, high=np.inf, shape=(self.window_days*390,25), dtype=np.float16)
         self.df = df
         # Every transcation to have this value ($)
         self.transaction_value = 1000
